feat: modularize Authlete MCP server architecture
- Split monolithic authlete_mcp_server.py (2,197 lines) into organized modules
- Create 6 specialized tool files organizing 36 MCP tools by functionality:
  * service_tools.py: 7 service management functions
  * client_tools.py: 16 client management functions
  * token_tools.py: 5 token operation functions
  * jose_tools.py: 2 JOSE operation functions
  * search_tools.py: 5 API/schema search functions
  * utility_tools.py: 1 JWKS generation function
- Maintain clean separation of concerns with dedicated modules:
  * config.py: Configuration and environment handling
  * models.py: Data models and schema definitions
  * api/: HTTP client logic for Authlete APIs
  * search.py: DuckDB-powered search engine
- Update main.py to use new modular server structure
- Update all tests to reference new server location
- All 35 unit tests continue to pass, ensuring full functionality preservation
- Remove legacy monolithic server file and update dependencies
- Update documentation to reflect new architecture
