+++ .githooks/pre-commit
@@ -48,7 +48,7 @@ uv sync --quiet
 
 # 1. Linting and Formatting Check
 print_status "Running linting checks..."
-if uv run ruff check authlete_mcp_server.py tests/ scripts/; then
+if uv run ruff check src/ tests/ scripts/ main.py; then
     print_success "Linting passed"
 else
     print_error "Linting failed. Run 'uv run ruff check --fix' to auto-fix some issues."
@@ -56,7 +56,7 @@ else
 fi
 
 print_status "Running format checks..."
-if uv run ruff format --check authlete_mcp_server.py tests/ scripts/; then
+if uv run ruff format --check src/ tests/ scripts/ main.py; then
     print_success "Code formatting is correct"
 else
     print_error "Code formatting failed. Run 'uv run ruff format' to fix formatting."
@@ -74,7 +74,7 @@ fi
 # 3. Type Checking (if mypy is available)
 print_status "Running type checks (if available)..."
 if uv run --quiet python -c "import mypy" 2>/dev/null; then
-    if uv run mypy authlete_mcp_server.py --ignore-missing-imports; then
+    if uv run mypy src/ main.py --ignore-missing-imports; then
         print_success "Type checking passed"
     else
         print_warning "Type checking issues found (non-blocking)"
