+++ main.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 """Main entry point for Authlete MCP Server."""
 
-from authlete_mcp_server import mcp
+from src.authlete_mcp.server import run
 
 if __name__ == "__main__":
-    mcp.run()
+    run()
