+++ src/authlete_mcp/__init__.py
@@ -1,5 +1,3 @@
-"""Authlete MCP Server - A Model Context Protocol server for Authlete API management."""
+"""Authlete MCP Server package."""
 
-__version__ = "1.0.0"
-__author__ = "Authlete, Inc."
-__description__ = "A Model Context Protocol server that provides tools for managing Authlete services and clients."
+__version__ = "0.1.0"
