+++ src/authlete_mcp/api/__init__.py
@@ -1,4 +1,4 @@
-"""Authlete API client modules."""
+"""API client for Authlete services."""
 
 from .client import make_authlete_idp_request, make_authlete_request
